+++ game/world/world.go
@@ -261,7 +261,7 @@ func (world *World) Render() {
 
 	// Setup 3D game render context
 	viewMat := camera.Transform.Matrix().Inv()
-	projMat := mgl32.Perspective(mgl32.DegToRad(settings.Current.Fov), settings.Current.WindowAspectRatio(), 0.1, 200.0)
+	projMat := mgl32.Perspective(mgl32.DegToRad(float32(settings.Current.Fov)), settings.Current.WindowAspectRatio(), 0.1, 200.0)
 	renderContext := render.Context{
 		View:           viewMat,
 		ViewInverse:    viewMat.Inv(),
